+++ store/client.ts
@@ -9,9 +9,13 @@ export type TableHint = {
   pine: string;
 };
 
-export type Hints = { table: TableHint[] };
+export type ColumnHint = {
+  column: string;
+};
+
+export type Hints = { table: TableHint[]; select: ColumnHint[] };
 // There are more operations. I'll add them as we need to handle them here
-export type Operation = { type: 'table' | 'delete' };
+export type Operation = { type: 'table' | 'delete' | 'select-partial' };
 export type Column = { column: string; alias: string };
 
 export type Ast = {
@@ -19,6 +23,7 @@ export type Ast = {
   'selected-tables': Table[];
   joins: string[][];
   context: string;
+  current: string;
   operation: Operation;
   columns: Column[];
 };